requirements.txt
Pillow==9.4.0
tweepy==4.12.1
